Refactor codebase with DI and separate files
